requirements.txt
streamlit
openai
rembg
pillow
requests